More accurate test fixture, split admin, and remove unused files
