+++ app/admin/__init__.py
@@ -1,0 +1,32 @@
+from ..models import Homepage, FAQ, History, Publication, Team, Leader, TeamMember, Collab, Location, Sponsor, Poster, Calendar, Video, Member
+from django.contrib.auth.models import User
+from django.contrib import admin
+from social_django.models import Association, Nonce, UserSocialAuth
+
+
+# Hide irrelevent Social Django models
+admin.site.unregister(Association)
+admin.site.unregister(Nonce)
+admin.site.unregister(UserSocialAuth)
+admin.site.unregister(User)
+
+# Register your models here.
+from .models import UserAdmin, ReportAdmin, SubteamAdmin
+admin.site.register(Homepage)
+admin.site.register(FAQ)
+admin.site.register(History)
+admin.site.register(Publication)
+admin.site.register(Team)
+admin.site.register(Sponsor)
+admin.site.register(TeamMember)
+admin.site.register(Poster)
+admin.site.register(Video)
+
+# WIP / Deprecated
+# admin.site.register(Calendar)
+# admin.site.register(Collab)
+# admin.site.register(Member)
+
+# Stagnant
+# admin.site.register(Leader)
+# admin.site.register(Location)
