+++ app/admin/models.py
@@ -1,11 +1,12 @@
-from .models import Homepage, FAQ, History, Publication, Team, Leader, TeamMember, RegisterUser, Collab, Location, Sponsor, Poster, ReportedCamera, Calendar, Video, Subteam, Member, CAM2dbApi
+from ..models import TeamMember, RegisterUser, ReportedCamera, Subteam, CAM2dbApi
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin #Important, dont remove
 from django.contrib.auth.models import User
-from .views.admin_actions import email_users, export_csv, report_csv, report_json
-from social_django.models import Association, Nonce, UserSocialAuth
+from .actions import email_users, export_csv, report_csv, report_json
+from social_django.models import UserSocialAuth
 
 # Override Django UserAdmin class
+@admin.register(User)
 class UserAdmin(UserAdmin):
     class UserInline(admin.StackedInline):
         model = RegisterUser
@@ -57,43 +58,13 @@ class UserAdmin(UserAdmin):
 
     country.short_description = 'country'
 
+@admin.register(ReportedCamera)
 class ReportAdmin(admin.ModelAdmin):
     actions = [report_csv, report_json]
 
+@admin.register(Subteam)
 class SubteamAdmin(admin.ModelAdmin):
     class MemberInline(admin.TabularInline):
         model = TeamMember
         extra = 0
     inlines = [MemberInline]
-
-
-# Register your models here.
-
-admin.site.unregister(User)
-admin.site.register(User, UserAdmin)
-
-admin.site.register(Homepage)
-admin.site.register(FAQ)
-admin.site.register(History)
-admin.site.register(Publication)
-admin.site.register(Team)
-admin.site.register(Sponsor)
-admin.site.register(ReportedCamera, ReportAdmin)
-admin.site.register(TeamMember)
-admin.site.register(Poster)
-admin.site.register(Video)
-admin.site.register(Subteam, SubteamAdmin)
-
-# WIP / Deprecated
-# admin.site.register(Calendar)
-# admin.site.register(Collab)
-# admin.site.register(Member)
-
-# Stagnant
-# admin.site.register(Leader)
-# admin.site.register(Location)
-
-# Hide irrelevent Social Django models
-admin.site.unregister(Association)
-admin.site.unregister(Nonce)
-admin.site.unregister(UserSocialAuth)
