--- app/templates/admin/base_html.html
@@ -1,1 +1,0 @@
-{% extends 'admin/base.html'}
