--- app/templates/templatetags/custom_tags.py
@@ -1,8 +1,0 @@
-from django import template
-
-register = template.Library()
-
-
-@register.filter(name='modulo')
-def modulo(value, arg):
-    return value % arg
